Include the SoftI2CMaster.h contents in TH02soft.
The problem is that SoftI2CMaster expects the SDA and SCL pins,
and some settings as well, to be set in #defines. But this makes it
pretty much impossible to have those definitions at one place only.
So now the whole SoftI2CMaster library is included here, and the
pin settings are done by editing TH02soft.cpp - still far from ideal,
having to edit the library files like this for change of pins.